requirements.txt
gradio==3.50.2
tensorflow
numpy
Pillow
